backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.1.0
bitarray==3.1.0
blinker==1.9.0
cachetools==5.3.1
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
cytoolz==1.0.1
dotenv==0.9.9
ecdsa==0.19.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
Flask==3.1.0
frozenlist==1.5.0
hexbytes==1.3.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
merkly==1.0.2
multidict==6.1.0
parsimonious==0.10.0
propcache==0.3.0
psycopg2-binary==2.9.10
pycryptodome==3.21.0
pydantic==1.10.21
pydantic_core==2.27.2
pymerkle==6.1.0
pysha3==1.0.2
python-dotenv==1.0.1
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
six==1.17.0
toolz==1.0.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
web3==7.8.0
websockets==13.1
Werkzeug==3.1.3
yarl==1.18.3
setuptools==75.8.2